Added mores tests to the CompositionSuite, and a bug in Composition was corrected.
The added tests found a bug in Composition of actors, and those were corrected: The transition from Processing to Idle just to process another message was introducing errors when there were concurrently messages being sent to the actor, and messages waiting to be processed. Now the ProxyActor only goes to idle after processing all the awaiting messages (while storing those that come in the middle of processing others).

Some minor edits to the ArrowOperator.